--- src/main/java/duke/task/TaskList.java
@@ -1,12 +1,0 @@
-package duke.task;
-
-import java.io.File;
-
-public class TaskList {
-
-    public static void main () {
-        File f = new File("data.txt");
-
-    }
-
-}
